backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
pandas
streamlit
plotly
